+++ packages/dynamic-flows/src/layout/button/spec.js
@@ -17,10 +17,10 @@ describe('Given a component for dynamically rendering buttons', () => {
         title: 'Submit',
         url: '/example',
         method: 'GET',
+        type: 'primary',
       },
       size: 'md',
       align: 'center',
-      context: 'primary',
       margin: 'md',
     };
     onAction = jest.fn();
@@ -29,6 +29,14 @@ describe('Given a component for dynamically rendering buttons', () => {
 
   it('should use the action title for the button', () => {
     expect(component.find(Button).contains(spec.action.title)).toBe(true);
+  });
+
+  it('should set correct type for button', () => {
+    expect(component.find(Button).prop('type')).toBe('accent');
+  });
+
+  it('should set correct priority for button', () => {
+    expect(component.find(Button).prop('priority')).toBe('primary');
   });
 
   describe('when the button is clicked', () => {