+++ packages/validation/src/index.js
@@ -1,4 +1,3 @@
-export * from './rule-validators';
 export * from './type-validators';
 export * from './value-validators';
 export { default as isKey } from './event-validators';